LPS-180956 Extract CommerceTaxMethodAvalaraScreenNavigationCategory methods to a CommerceTaxMethodAvalaraScreenNavigationEntry component and extend CommerceTaxMethodAvalaraScreenNavigationCategory in it
